refactor(scripts): extract git ref resolution into `lib/git.js`
Move the remote lookup, `fetchAndResolve()` and the pull request number
shorthand out of `diff-flat.js`, so that other scripts can resolve the
same base and head references.
